implement RetargetingHlient library
