+++ src/CMakeLists.txt
@@ -11,3 +11,4 @@ add_subdirectory(libRobot_helper)
 add_subdirectory(libWhole_body_controllers)
 add_subdirectory(libTrajectory_planner)
 add_subdirectory(libKinDyn_wrapper)
+add_subdirectory(libRetargeting_helper)
